[REFACTOR] Update Tab component imports and remove unused Tab files
- Changed Tab component import paths to a centralized location.
- Deleted unused Tab component files and their associated types to clean up the codebase.
