+++ src/components/Tab/types.ts
